Refactor: 웹소켓 연결 코드 커스텀 훅 분리
