+++ frontend/src/components/chat/ChatMessage.tsx
@@ -8,7 +8,7 @@ interface ChatMessageListProps {
 
 const ChatMessage = ({ messages, username }: ChatMessageListProps) => {
 	let currentFormattedDate = '';
-	console.log('메시지', username);
+	// console.log('본인', username);
 
 	return (
 		<>
@@ -37,12 +37,22 @@ const ChatMessage = ({ messages, username }: ChatMessageListProps) => {
 
 				currentFormattedDate = formattedDate;
 
+				console.log('boolean', item.sender === username);
+				console.log(item.sender);
+
 				return (
 					<MessageContainer key={index}>
 						{shouldDisplayYear && <DateWrapper>{formattedDate}</DateWrapper>}
 						<MessageWrapper>
 							{/* 추후 코드 본인일 경우 상태관리 추가해야됨 */}
 							{item.sender === username ? (
+								<ChatReverseWrapper>
+									<ChatOwnBox>{item.content}</ChatOwnBox>
+									<ChatDateWrapper>
+										<FormattedTime>{isLastMessageForWriter && formattedTime}</FormattedTime>
+									</ChatDateWrapper>
+								</ChatReverseWrapper>
+							) : (
 								<>
 									<ChatSenderWrapper>
 										<ChatSender>{item.sender}</ChatSender>
@@ -54,13 +64,6 @@ const ChatMessage = ({ messages, username }: ChatMessageListProps) => {
 										</ChatDateWrapper>
 									</ChatWrapper>
 								</>
-							) : (
-								<ChatReverseWrapper>
-									<ChatOwnBox>{item.content}</ChatOwnBox>
-									<ChatDateWrapper>
-										<FormattedTime>{isLastMessageForWriter && formattedTime}</FormattedTime>
-									</ChatDateWrapper>
-								</ChatReverseWrapper>
 							)}
 						</MessageWrapper>
 					</MessageContainer>
